pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hostsolo"
version = "0.1.0"
description = "Lightweight self-hosted application platform for VPS deployments"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["docker", "vps", "self-hosted", "deployment", "traefik"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "typer[all]>=0.9.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "docker>=7.0",
    "boto3>=1.34",
    "httpx>=0.27",
    "jinja2>=3.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.3.0",
    "mypy>=1.8",
    "types-PyYAML",
    "boto3-stubs[s3]",
]

[project.scripts]
hostsolo = "hostsolo.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/host-solo"
Repository = "https://github.com/yourusername/host-solo"

[tool.hatch.build.targets.wheel]
packages = ["hostsolo"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=hostsolo"
